Add media query hook and update RecentPostsWidget visibility logic
