+++ src/components/RecentPostsWidget.tsx
@@ -5,6 +5,8 @@ import { formatPublishedDate } from '~/utils/blog'
 
 type RecentPostsWidgetProps = {
   posts?: ReadonlyArray<RecentPost>
+  /** Set to false to skip the client fetch when the widget is rendered but not visible (e.g. hidden below a CSS breakpoint). Ignored when `posts` is provided. */
+  enabled?: boolean
 }
 
 function RecentPostsList({ posts }: { posts: ReadonlyArray<RecentPost> }) {
@@ -63,11 +65,14 @@ function RecentPostsSkeleton() {
   )
 }
 
-export function RecentPostsWidget({ posts }: RecentPostsWidgetProps) {
+export function RecentPostsWidget({
+  posts,
+  enabled = true,
+}: RecentPostsWidgetProps) {
   const recentPostsQuery = useQuery({
     queryKey: ['recentPosts'],
     queryFn: () => fetchRecentPosts(),
-    enabled: posts === undefined,
+    enabled: posts === undefined && enabled,
     staleTime: 1000 * 60 * 5,
   })
