+++ src/routes/blog.$.tsx
@@ -9,6 +9,7 @@ import { LibrariesWidget } from '~/components/LibrariesWidget'
 import { partners } from '~/utils/partners'
 import { PartnersRail, RightRail } from '~/components/RightRail'
 import { RecentPostsWidget } from '~/components/RecentPostsWidget'
+import { useMediaQuery } from '~/utils/useMediaQuery'
 
 import { Toc } from '~/components/Toc'
 import { Breadcrumbs } from '~/components/Breadcrumbs'
@@ -74,6 +75,7 @@ function BlogPost() {
   const { _splat: slug } = Route.useParams()
   const markdown = React.useMemo(() => parseSiteMarkdown(content), [content])
   const headings = markdown.headings
+  const isDesktopViewport = useMediaQuery('(min-width: 768px)')
 
   const isTocVisible = headings.length > 1
 
@@ -203,7 +205,7 @@ function BlogPost() {
               partners={activePartners}
             />
             <div className="hidden md:block border border-gray-500/20 rounded-l-lg overflow-hidden w-full">
-              <RecentPostsWidget />
+              <RecentPostsWidget enabled={isDesktopViewport} />
             </div>
             <Card>
               <LibrariesWidget />
