+++ src/utils/useMediaQuery.ts
@@ -1,0 +1,21 @@
+import * as React from 'react'
+
+export function useMediaQuery(query: string) {
+  const [matches, setMatches] = React.useState(false)
+
+  React.useEffect(() => {
+    const mediaQueryList = window.matchMedia(query)
+    const updateMatches = () => {
+      setMatches(mediaQueryList.matches)
+    }
+
+    updateMatches()
+    mediaQueryList.addEventListener('change', updateMatches)
+
+    return () => {
+      mediaQueryList.removeEventListener('change', updateMatches)
+    }
+  }, [query])
+
+  return matches
+}
